requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
replicate==0.22.0
requests==2.31.0
Werkzeug==3.0.1
python-magic==0.4.27
redis==5.0.1
rq==1.15.1
Flask-Mail==0.9.1
gunicorn==21.2.0
boto3==1.34.0
